Add files via upload
This is the final project to code along to in the course and have competed the Bootstrap course on Udemy created by Brad Traversy.  The portfolio website can be improved now using these new skills and it'll help when creating apps using the MEAN stack, namely the dating app that has been put on the backburner for longer than wanted.